+++ eventsourcing/application/system.py
@@ -5,7 +5,6 @@ from queue import Empty, Queue
 from threading import Barrier, BrokenBarrierError, Event, Lock, Thread, Timer
 from time import sleep
 
-from eventsourcing.application.notificationlog import NotificationLogReader
 from eventsourcing.application.popo import PopoApplication
 from eventsourcing.application.process import ProcessApplication, Prompt
 from eventsourcing.application.simple import ApplicationWithConcreteInfrastructure
@@ -44,6 +43,7 @@ class System(object):
         self.pipelines_exprs = pipeline_exprs
         self.setup_tables = kwargs.get('setup_tables', False)
         self.infrastructure_class = kwargs.get('infrastructure_class', None)
+        self.use_direct_query_if_available = kwargs.get('use_direct_query_if_available', False)
 
         self.session = kwargs.get('session', None)
 
@@ -113,7 +113,10 @@ class System(object):
         """
         Supports usage of a system object as a context manager.
         """
-        self.__runner = SingleThreadedRunner(self)
+        self.__runner = SingleThreadedRunner(
+            self,
+            use_direct_query_if_available=self.use_direct_query_if_available
+        )
         self.__runner.__enter__()
         return self
 
@@ -144,7 +147,8 @@ class System(object):
 
 class SystemRunner(ABC):
 
-    def __init__(self, system: System, infrastructure_class=None, setup_tables=False):
+    def __init__(self, system: System, infrastructure_class=None, setup_tables=False,
+                 use_direct_query_if_available=False):
         self.system = system
         self.infrastructure_class = infrastructure_class or self.system.infrastructure_class
         # Check that a concrete infrastructure class is involved.
@@ -155,6 +159,8 @@ class SystemRunner(ABC):
             ):
                 raise ProgrammingError("System runner needs a concrete application infrastructure class")
         self.setup_tables = setup_tables
+        self.use_direct_query_if_available = use_direct_query_if_available or \
+                                             system.use_direct_query_if_available
         self.processes = {}
 
     def __enter__(self):
@@ -198,6 +204,7 @@ class InProcessRunner(SystemRunner):
                 process_class=process_class,
                 infrastructure_class=self.infrastructure_class,
                 setup_table=self.setup_tables or self.system.setup_tables,
+                use_direct_query_if_available=self.use_direct_query_if_available,
             )
             self.system.processes[process.name] = process
 
@@ -552,7 +559,8 @@ class SteppingSingleThreadedRunner(SteppingRunner):
             stop_event=self.stop_event,
             is_verbose=self.is_verbose,
             seen_prompt_events=self.seen_prompt_events,
-            processes=self.system.processes
+            processes=self.system.processes,
+            use_direct_query_if_available=self.use_direct_query_if_available
         )
         self.clock_thread.start()
 
@@ -566,7 +574,7 @@ class SteppingSingleThreadedRunner(SteppingRunner):
         super(SteppingSingleThreadedRunner, self).close()
         self.clock_thread.join(5)
         if self.clock_thread.isAlive():
-            print(f"Warning: clock thread was still alive")
+            print("Warning: clock thread was still alive")
 
 
 class ClockThread(Thread):
@@ -586,13 +594,15 @@ class ClockThread(Thread):
 
 class ProcessRunningClockThread(ClockThread):
     def __init__(self, normal_speed, scale_factor, stop_event: Event,
-                 is_verbose=False, seen_prompt_events=None, processes=None):
+                 is_verbose=False, seen_prompt_events=None, processes=None,
+                 use_direct_query_if_available=False):
         super(ProcessRunningClockThread, self).__init__(daemon=True)
         self.normal_speed = normal_speed
         self.scale_factor = scale_factor
         self.stop_event = stop_event
         self.seen_prompt_events = seen_prompt_events
         self.processes = processes
+        self.use_direct_query_if_available = use_direct_query_if_available
         self.last_tick_time = None
         self.last_process_time = None
         self.all_tick_durations = deque()
@@ -618,9 +628,9 @@ class ProcessRunningClockThread(ClockThread):
         # Construct a notification log reader for each process.
         self.readers = {}
         for process_name, process in self.processes.items():
-            reader = NotificationLogReader(
+            reader = process.notification_log_reader_class(
                 notification_log=process.notification_log,
-                use_direct_query_if_available=True
+                use_direct_query_if_available=self.use_direct_query_if_available
             )
             self.readers[process_name] = reader